requirements.txt
asgiref==3.3.4
click==7.1.2
Django==3.2.2
django-autoslug==1.9.8
django-crispy-forms==1.11.2
gunicorn==20.1.0
pep517==0.10.0
Pillow==8.2.0
pip-tools==6.1.0
pkg-resources==0.0.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
